+++ services/meta/data.go
@@ -1363,9 +1363,16 @@ func (sgi ShardGroupInfo) clone() ShardGroupInfo {
 	return other
 }
 
-// ShardFor returns the ShardInfo for a Point hash.
-func (sgi *ShardGroupInfo) ShardFor(hash uint64) ShardInfo {
-	return sgi.Shards[hash%uint64(len(sgi.Shards))]
+type hashIDer interface {
+	HashID() uint64
+}
+
+// ShardFor returns the ShardInfo for a Point or other hashIDer.
+func (sgi *ShardGroupInfo) ShardFor(p hashIDer) ShardInfo {
+	if len(sgi.Shards) == 1 {
+		return sgi.Shards[0]
+	}
+	return sgi.Shards[p.HashID()%uint64(len(sgi.Shards))]
 }
 
 // marshal serializes to a protobuf representation.